Generate TIDs in the library
Azalea doesn't have 'thread IDs' as such - it has handles, which are
64-bits long, and multiple handles may refer to the same thread, whereas
large chunks of musl are built on 32-bit TIDs that are unique per
thread. It will be quite a lot of effort to give Azalea the same
concept, and I'm not sure it's needed. Also, to make musl use 64-bit
TIDs is quite a lot of effort - it interacts with all kinds of things
like locks and futexes.

Instead, add a simple 'thread ID manager' to the library. All it does
right now is hand out monotonically increasing IDs. If any application
uses more than 2^31 threads in its lifetime, I'll update this.